\name{EnVarIP}
\docType{data}
\alias{EnVarIP}
\encoding{latin1}
\title{ENVIRONMENTAL VARIABLES IN THE IBERIAN PENINSULA}
\usage{data(EnVarIP)}
\description{Geographic coordinates with values of temperature and precipitation in the Iberian Peninsula.} 
\format{Data frame with longitude, latitude and values of temperature, rainfall and altitude, and the geographic coordinates of the polygons of Spain and Portugal in the Iberian Peninsula.}
\keyword{EnVarIP}
