\name{Rotifers}
\docType{data}
\alias{Rotifers}
\encoding{latin1}
\title{ABUNDANCE OF ROTIFERS IN PONDS}
\usage{data(Rotifers)}
\description{Abundance of species of rotifers in ponds of Doñana National Park (Spain), which were obtained from Table 1 of Mazuelos et al. (1993).} 
\format{An data frame with 40 columns: position of the label of the pond in the DER plot, 7 columns with the taxonomy of the species and the rest of columns are the abundance of the species of rotifers in the ponds.}
\references{
Mazuelos, N., Toja, J., & Guisande, C. (1993) Rotifers in ephemeral ponds of Doñana National Park. \emph{Hidrobiologia}, 255/256: 429-434.
}
\keyword{Rotifers}
